fix: remove fcntl func add ioctl func
